refactor: 타임라인 패키지 변경(#178)
