+++ presentation/src/main/java/com/example/smiley/main/home/adapter/timeline/TimeLineItem.kt
@@ -1,4 +1,4 @@
-package com.example.smiley.main.home.adapter
+package com.example.smiley.main.home.adapter.timeline
 
 data class TimeLineItem(
     val viewType: String,
